% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{credit}
\alias{credit}
\title{Credit scoring data.}
\format{A data frame with 10 rows and 6 variables:
\describe{
  \item{age}{age of person}
  \item{married}{is the person married or not, boolean}
  \item{house}{is the person a homeowner or not, boolean}
  \item{income}{income of person, in thousands}
  \item{gender}{factor with levels male, female}
  \item{class}{class variable (0 or 1)}
}}
\usage{
credit
}
\description{
A dataset containing the attributes of people who did or did not default on their loan.
}
\details{
Toy example, useful for debugging purposes.
}
\keyword{datasets}
