% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{peelCpp}
\alias{peelCpp}
\title{PRIM Peel}
\usage{
peelCpp(M, y, colTypes, alpha, minSup)
}
\arguments{
\item{M}{The data to peel away from}

\item{y}{The labels to evaluate peels}

\item{colTypes}{Indicates which columns are numeric (0) and which are categorical (1)}

\item{alpha}{The peeling quantile}

\item{minSup}{The minimum allowed size of the remainder after a peel}
}
\value{
A list of peeling steps
}
\description{
This function iteratively peels away a small portion of the matrix M
such that the mean of the remaining values in y is maximized.
}
\author{
Jurian Baas
}
