% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{plot.prim.cover}
\alias{plot.prim.cover}
\title{Plot PRIM cover result}
\usage{
\method{plot}{prim.cover}(x, ...)
}
\arguments{
\item{x}{An S3 object of class prim.cover}

\item{...}{Optional arguments to pass on}
}
\value{
Nothing, this function is called for its side-effects
}
\description{
Plot an S3 object of class prim.cover
}
\author{
Jurian Baas
}
