% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{plot.prim.diversify}
\alias{plot.prim.diversify}
\title{Plot PRIM diversify result}
\usage{
\method{plot}{prim.diversify}(x, ...)
}
\arguments{
\item{x}{An S3 object of class prim.diversify}

\item{...}{Optional arguments to pass on}
}
\value{
Nothing, this function is called for its side-effects
}
\description{
Plot an S3 object of class prim.diversify
}
\author{
Jurian Baas
}
