% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{plot.prim.peel}
\alias{plot.prim.peel}
\title{Plot PRIM peel result}
\usage{
\method{plot}{prim.peel}(x, ...)
}
\arguments{
\item{x}{An S3 object of class prim.peel}

\item{...}{Optional arguments to pass on}
}
\value{
Nothing, this function is called for its side-effects
}
\description{
Plot an S3 object of class prim.peel
}
\author{
Jurian Baas
}
