% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{plot.prim.peel.result}
\alias{plot.prim.peel.result}
\title{Plot PRIM peel result}
\usage{
\method{plot}{prim.peel.result}(x, ...)
}
\arguments{
\item{x}{An S3 object of class prim.peel.result}

\item{...}{Optional arguments to pass on}
}
\description{
Plot an S3 object of class prim.peel.result
}
\author{
Jurian Baas
}
