% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.functions.R
\name{plot.prim.test.result}
\alias{plot.prim.test.result}
\title{Plot PRIM test result}
\usage{
\method{plot}{prim.test.result}(test.result)
}
\arguments{
\item{test.result}{An S3 object of class prim.test.result}
}
\description{
Plot an S3 object of class prim.test.result
}
\author{
Jurian Baas
}
