% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{plot.prim.validate}
\alias{plot.prim.validate}
\title{Plot PRIM test result}
\usage{
\method{plot}{prim.validate}(x, ...)
}
\arguments{
\item{x}{An S3 object of class prim.validate}

\item{...}{Optional arguments to pass on}
}
\value{
Nothing, this function is called for its side-effects
}
\description{
Plot an S3 object of class prim.validate
}
\author{
Jurian Baas
}
