% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{predict.prim.cover}
\alias{predict.prim.cover}
\title{Predict method for PRIM Cover Fits}
\usage{
\method{predict}{prim.cover}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class prim.cover}

\item{newdata}{A data frame in which to look for variables with which to predict.}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Depends on the quality function used. In the case of base::mean, the mean of the target variable of the first matching box.
}
\description{
Predicted values based on the PRIM cover object
}
\author{
Jurian Baas
}
