% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.box.optimal}
\alias{prim.box.optimal}
\title{Find the optimal box depending on the strategy}
\usage{
prim.box.optimal(prim.predict)
}
\arguments{
\item{prim.predict}{An object of type "prim.predict"}
}
\value{
The index of the optimal box
}
\description{
Finds the box with the highest quality or the box closest to the maximum quality minus 2 times the standard error
}
\author{
Jurian Baas
}
