% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.functions.R
\name{prim.candidates.best}
\alias{prim.candidates.best}
\title{Find optimal box candidate}
\usage{
prim.candidates.best(candidates)
}
\arguments{
\item{candidates}{List of candidate generated by prim.candidates.find()}
}
\value{
A list with the optimal candidate information
}
\description{
This function goes through the box candidate qualities and finds the optimal candidate
}
\author{
Jurian Baas
}
