% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.functions.R
\name{prim.candidates.find}
\alias{prim.candidates.find}
\title{Find all box candidates for a given (sub)set}
\usage{
prim.candidates.find(X, alpha, min.support, N)
}
\arguments{
\item{X}{Data frame with observations (may be a subset of original data)}

\item{alpha}{Quantile to peel off}

\item{min.support}{Minimal size of a box}

\item{N}{Number of rows of original subset}
}
\value{
A list of potential boxes
}
\description{
This function goes through all columns of the dataset and tries to findbox candidates based on the quantile alpha and minimum support min.support. Note that the indexes returned are those that have to be removed in order to create the box!
}
\author{
Jurian Baas
}
