% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.functions.R
\name{prim.candidates.quality}
\alias{prim.candidates.quality}
\title{Candidate box quality}
\usage{
prim.candidates.quality(y, candidates, quality.function)
}
\arguments{
\item{y}{Vector of values on which to determine the quality}

\item{candidates}{List of candidates generated by prim.candidates.find()}

\item{quality.function}{Function to use for a quality measure (usually mean)}
}
\value{
A list with the qualities of each supplied candidate box
}
\description{
This function calculates the quality of the dependant variable y given some box candidate
}
\author{
Jurian Baas
}
