% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.cover.formula}
\alias{prim.cover.formula}
\title{PRIM covering algorithm}
\usage{
prim.cover.formula(formula, data, peeling.quantile, min.support,
  train.fraction = 0.66, max.boxes = NA, quality.function = mean)
}
\arguments{
\item{formula}{Formula with a response and terms}

\item{data}{Data frame to find rules in}

\item{peeling.quantile}{Quantile to peel off for numerical variables}

\item{min.support}{Minimal size of a box to be valid}

\item{train.fraction}{Train-test split fraction used in validation}

\item{max.boxes}{Optional maximum number of boxes}

\item{quality.function}{Optional setting for function to use for determining subset quality, defaults to mean}
}
\description{
In bump hunting it is customary to follow a so-called covering strategy. This means that the same box construction (rule induction) algorithm is applied sequentially to subsets of the data.
}
\author{
Jurian Baas
}
