% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.cover}
\alias{prim.cover}
\title{PRIM covering algorithm}
\usage{
prim.cover(X, y, peeling.quantile, min.support, train.fraction = 0.66,
  max.boxes = NA, quality.function = mean)
}
\arguments{
\item{X}{Data frame to find rules in}

\item{y}{Response vector, usually of type numeric}

\item{peeling.quantile}{Quantile to peel off for numerical variables}

\item{min.support}{Minimal size of a box to be valid}

\item{train.fraction}{Train-test split fraction used in validation}

\item{max.boxes}{Optional maximum number of boxes}

\item{quality.function}{Optional setting for function to use for determining subset quality, defaults to mean}
}
\description{
Temporary description
}
\author{
Jurian Baas
}
