% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.diversify.compare}
\alias{prim.diversify.compare}
\title{Compare PRIM diversify results}
\usage{
prim.diversify.compare(X, p.div)
}
\arguments{
\item{X}{Data frame to do intersect and union operations}

\item{p.div}{An S3 object of type "prim.diversify"}
}
\value{
A matrix with the comparisons laid out by position
}
\description{
Compares all attempts of a PRIM diversify operation with each other
}
\details{
Comparison is done by the following formula \eqn{\frac{|A \cap B|}{|A \cup B|}}
}
\author{
Jurian Baas
}
