% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.diversify}
\alias{prim.diversify}
\title{PRIM main entrance}
\usage{
prim.diversify(formula, data, X = NULL, y = NULL, n, peeling.quantile,
  min.support, train.fraction = 0.66, quality.function = mean)
}
\arguments{
\item{formula}{Formula with a response and terms}

\item{data}{Data frame to find rules in}

\item{X}{Optionally instead of using a formula: Data frame to find rules in}

\item{y}{Optionally instead of using a formula: Response vector, usually of type numeric}

\item{n}{Numer of attempts to run the PRIM algorithm}

\item{peeling.quantile}{Quantile to peel off for numerical variables}

\item{min.support}{Minimal size of a box to be valid}

\item{train.fraction}{Optional train-test split fraction used in validation, defaults to 0.66}

\item{quality.function}{Optional setting for function to use for determining subset quality, defaults to mean}
}
\description{
Temporary description
}
\author{
Jurian Baas
}
