% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.functions.R
\name{prim.formula}
\alias{prim.formula}
\title{Bump hunting using the Patient Rule Induction Method}
\usage{
prim.formula(formula, data, alpha, min.support, quality.function = mean)
}
\arguments{
\item{formula}{Formula with a response and terms}

\item{data}{Data frame to find rules in}

\item{alpha}{Quantile to peel off for numerical variables}

\item{min.support}{Minimal size of a box to be valid}

\item{quality.function}{Which function to use to determine the quality of a box, defaults to mean}
}
\value{
An S3 object of class prim.peel.result
}
\description{
Main function for bump hunting using the Patient Rule Induction Method (PRIM). N.B. Remember to remove the intercept from the formula!
}
\author{
Jurian Baas
}
