% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.match.rule}
\alias{prim.match.rule}
\title{Create box matching index}
\usage{
prim.match.rule(prim.object, X)
}
\arguments{
\item{prim.object}{An S3 object of class prim.peel or prim.validate result}

\item{X}{A data frame with at least those columns that were used in creating the prim S3 object}
}
\value{
A logical index of matching records
}
\description{
Generate a logical vector in which elements are true iff applying the rule to a record evaluates to true.
}
\author{
Jurian Baas
}
