% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.peel}
\alias{prim.peel}
\title{Find the optimal sub-box using the PRIM peeling strategy}
\usage{
prim.peel(
  formula,
  data,
  X,
  y,
  peeling.quantile = 0.03,
  min.support = 0.05,
  parallel = T
)
}
\arguments{
\item{formula}{Formula with a response and terms}

\item{data}{Data frame to find rules in}

\item{X}{Optionally instead of using a formula: Data frame to find rules in}

\item{y}{Optionally instead of using a formula: Response vector, usually of type numeric}

\item{peeling.quantile}{Quantile to peel off for numerical variables}

\item{min.support}{Minimal size of a box to be valid}

\item{parallel}{Peel columns in parallel, faster for many columns (defaults to true)}
}
\value{
An S3 object of class prim.peel
}
\description{
By iteratively removing a small portion of the data... Note that categorical data has to be in factor form.
}
\author{
Jurian Baas
}
