% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.predict.metrics}
\alias{prim.predict.metrics}
\title{Calculate statistical metrics}
\usage{
prim.predict.metrics(prim.predict)
}
\arguments{
\item{prim.predict}{An object of type "prim validate"}
}
\value{
A list with elements described above
}
\description{
This function calculates the mean, standard deviation, standard error of the mean, 95% confidence intervals
}
\author{
Jurian Baas
}
