% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.functions.R
\name{prim}
\alias{prim}
\title{Main function for bump hunting using the Patient Rule Induction Method (PRIM).}
\usage{
prim(X, y, alpha, min.support, quality.function = mean)
}
\arguments{
\item{X}{Data frame to find rules in}

\item{y}{Response vector, usually of type numeric}

\item{alpha}{Quantile to peel off for numerical variables}

\item{min.support}{Minimal size of a box to be valid}

\item{quality.funciton}{Which function to use to determine the quality of a box, defaults to mean}
}
\value{
An S3 object of class prim.peel.result
}
\description{
Main function for bump hunting using the Patient Rule Induction Method (PRIM).
}
\author{
Jurian Baas
}
