% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.rule.condense}
\alias{prim.rule.condense}
\title{Condense multiple (redundant) rules}
\usage{
prim.rule.condense(prim.object)
}
\arguments{
\item{prim.object}{An S3 object of class prim.peel or prim.validate}
}
\value{
The condensed rule as a single string
}
\description{
This function condenses the many (redundant) rules of an S3 object of class prim.peel or prim.validate to a single rule.
}
\details{
This function condenses the many (redundant) rules of an S3 object of class prim.peel or prim.validate to a single rule.
}
\author{
Jurian Baas
}
