% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.rule.intersect}
\alias{prim.rule.intersect}
\title{Intersection of multiple rules}
\usage{
prim.rule.intersect(X, ...)
}
\arguments{
\item{X}{Data frame to apply rules to}

\item{...}{A number of objects of class "prim.peel" and/or "prim.validate"}
}
\value{
Logical vector, true iff all rules evaluate to TRUE for a certain observation
}
\description{
This function applies the rules given by the parameters to a dataset and calculates the intersection, i.e. those observations where all rules evaluate to TRUE are returned as TRUE.
}
\author{
Jurian Baas
}
