% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.rule.union}
\alias{prim.rule.union}
\title{Union of multiple rules}
\usage{
prim.rule.union(X, ...)
}
\arguments{
\item{X}{Data frame to apply rules to}

\item{...}{A number of objects of class "prim.peel" and/or "prim.validate"}
}
\value{
Logical vector, true iff at least one rule evalutes to TRUE a certain observation
}
\description{
This function applies the rules given by the parameters to a dataset and calculates the union, i.e. those observations where at least one rule evaluates to TRUE are returned as TRUE.
}
\author{
Jurian Baas
}
