% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.functions.R
\name{prim.subset.box}
\alias{prim.subset.box}
\title{PRIM subset creator}
\usage{
prim.subset.box(prim.object, X)
}
\arguments{
\item{prim.object}{An S3 object of class prim.peel.result or prim.test result}

\item{X}{A data frame with at least those columns that were used in creating the prim S3 object}
}
\value{
A subset of X
}
\description{
Generate a subset of the data using the rules in the supplied prim S3 object
}
\author{
Jurian Baas
}
