% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.functions.R
\name{prim.test}
\alias{prim.test}
\title{Bump hunting using the Patient Rule Induction Method}
\usage{
prim.test(peel.result, X, y)
}
\arguments{
\item{peel.result}{An S3 object of class prim.peel.result}

\item{X}{A data frame with at least those columns that were used in creating the prim.peel.result S3 object}

\item{y}{Response vector, usually of type numeric}
}
\value{
An S3 object of type prim.test.result
}
\description{
This function takes the result of prim() and applies it to new data. Usually the optimal box in the peeling process is not the best on unobserved data.
}
\author{
Jurian Baas
}
