% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.validate.metrics}
\alias{prim.validate.metrics}
\title{Calculate statistical metrics}
\usage{
prim.validate.metrics(prim.validate)
}
\arguments{
\item{prim.validate}{An object of type "prim validate"}
}
\value{
A list with elements described above
}
\description{
This function calculates the mean, standard deviation, standard error of the mean, 95% confidence intervals
}
\author{
Jurian Baas
}
