% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{quasi.convex.hull}
\alias{quasi.convex.hull}
\title{Calculate a frontier of dominating points}
\usage{
quasi.convex.hull(p.div)
}
\arguments{
\item{p.div}{An object of type "prim.diversify"}
}
\value{
A vector of indexes for the dominating points
}
\description{
During the diversify process, we are really only interested in the attempts which dominate all others in performance.
}
\seealso{
\url{https://stats.stackexchange.com/a/65157}
}
\author{
William Huber & Jurian Baas
}
