% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simplifyRules}
\alias{simplifyRules}
\title{PRIM Simplify Rules}
\usage{
simplifyRules(peelSteps, colTypes, bestBoxIndex)
}
\arguments{
\item{peelSteps}{A list of peeling steps}

\item{colTypes}{Indicates which columns are numeric (0) and which are categorical (1)}

\item{bestBoxIndex}{The (zero) index of the best box}
}
\value{
A list of peeling steps
}
\description{
This function will go through all boxes that were found and tries to remove redundant ones,
as well as grouping them by column
}
\author{
Jurian Baas
}
