% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{summary.prim.peel}
\alias{summary.prim.peel}
\title{Summarize a PRIM peeling result object}
\usage{
\method{summary}{prim.peel}(object, ...)
}
\arguments{
\item{object}{An S3 object of class prim.peel}

\item{...}{Optional arguments to pass on}
}
\description{
Summarize a PRIM peeling result object
}
\author{
Jurian Baas
}
