% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{summary.prim.test.result}
\alias{summary.prim.test.result}
\title{Summarize a PRIM test result object}
\usage{
\method{summary}{prim.test.result}(object, ...)
}
\arguments{
\item{object}{An S3 object of class prim.test.result}

\item{...}{Optional arguments to pass on}
}
\description{
Summarize a PRIM test result object
}
\author{
Jurian Baas
}
