% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate}
\alias{validate}
\title{PRIM Validate}
\usage{
validate(peelSteps, M, y)
}
\arguments{
\item{peelSteps}{Peeling result from calling peel()}

\item{M}{The data to peel away from}

\item{y}{The labels to evaluate peels}
}
\value{
A list of peeling steps
}
\description{
This function iteratively goes through the steps from the peeling process
and removes any trailing peels which do not improve the overall quality.
}
\author{
Jurian Baas
}
