% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.functions.R
\name{plot.prim.peel.result}
\alias{plot.prim.peel.result}
\title{Plot an S3 object of class prim.peel.result}
\usage{
\method{plot}{prim.peel.result}(peel.result)
}
\arguments{
\item{peel.result}{An S3 object of class prim.peel.result}
}
\description{
Plot an S3 object of class prim.peel.result
}
\author{
Jurian Baas
}
