% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{plot.prim.test.result}
\alias{plot.prim.test.result}
\title{Plot PRIM test result}
\usage{
\method{plot}{prim.test.result}(x, ...)
}
\arguments{
\item{x}{An S3 object of class prim.test.result}

\item{...}{Optional arguments to pass on}
}
\description{
Plot an S3 object of class prim.test.result
}
\author{
Jurian Baas
}
