% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.functions.R
\name{prim.candidates.best}
\alias{prim.candidates.best}
\title{Find optimal box candidate}
\usage{
prim.candidates.best(candidates.quality)
}
\arguments{
\item{candidates.quality}{List of candidate qualities generated by prim.candidates.quality()}
}
\value{
Index of the optimal candidate in the candidates.quality argument
}
\description{
This function goes through the box candidate qualities and finds the optimal candidate
}
\author{
Jurian Baas
}
