% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.functions.R
\name{prim.condense.rules}
\alias{prim.condense.rules}
\title{Condense multiple (redundant) rules}
\usage{
prim.condense.rules(prim.object)
}
\arguments{
\item{prim.object}{An S3 object of class prim.peel.result or prim.test.result}
}
\value{
The condensed rule as a single string
}
\description{
This function condenses the many (redundant) rules of an S3 object of class prim.peel.result or prim.test.result to a single rule.
}
\author{
Jurian Baas
}
