% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.cover}
\alias{prim.cover}
\title{PRIM covering strategy}
\usage{
prim.cover(formula, data, X, y, peeling.quantile, min.support, max.peel = 0.1,
  train.fraction = 0.66, max.boxes = NA, quality.function = base::mean,
  plot = FALSE)
}
\arguments{
\item{formula}{Formula with a response and terms}

\item{data}{Data frame to find rules in}

\item{X}{Optionally instead of using a formula: Data frame to find rules in}

\item{y}{Optionally instead of using a formula: Response vector, usually of type numeric}

\item{peeling.quantile}{Quantile to peel off for numerical variables}

\item{min.support}{Minimal size of a box to be valid}

\item{max.peel}{Maximal size of a peel, as a fraction. Defaults to 0.1}

\item{train.fraction}{Train-test split fraction used in validation, defaults to 0.66}

\item{max.boxes}{Optional maximum number of boxes}

\item{quality.function}{Optional setting for function to use for determining subset quality, defaults to mean}

\item{plot}{Optional setting to plot intermediate results, defaults to false}
}
\value{
An S3 object of class prim.cover
}
\description{
In bump hunting it is customary to follow a so-called covering strategy. This means that the same box construction (rule induction) algorithm is applied sequentially to subsets of the data.
}
\author{
Jurian Baas
}
