% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.peel}
\alias{prim.peel}
\title{Bump hunting using the Patient Rule Induction Method}
\usage{
prim.peel(formula, data, X = NULL, y = NULL, peeling.quantile, min.support,
  quality.function = mean)
}
\arguments{
\item{formula}{Formula with a response and terms}

\item{data}{Data frame to find rules in}

\item{X}{Data frame to find rules in}

\item{y}{Response vector, usually of type numeric}

\item{peeling.quantile}{Quantile to peel off for numerical variables}

\item{min.support}{Minimal size of a box to be valid, as a fraction}

\item{quality.function}{Which function to use to determine the quality of a box, defaults to mean}
}
\value{
An S3 object of class prim.peel
}
\description{
Peeling function for bump hunting using the Patient Rule Induction Method (PRIM).
}
\author{
Jurian Baas
}
