% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{summary.prim.cover}
\alias{summary.prim.cover}
\title{Summarize a PRIM cover result object}
\usage{
\method{summary}{prim.cover}(object, ...)
}
\arguments{
\item{object}{An S3 object of class prim.cover}

\item{...}{Optional arguments to pass on}
}
\description{
Summarize a PRIM cover result object
}
\author{
Jurian Baas
}
