% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{summary.prim.diversify}
\alias{summary.prim.diversify}
\title{Summarize a PRIM cover diversify object}
\usage{
\method{summary}{prim.diversify}(object, ..., round = TRUE, digits = 2)
}
\arguments{
\item{object}{An S3 object of class prim.diversify}

\item{...}{Optional arguments to pass on}

\item{round}{Optional setting to disable rounding}

\item{digits}{Optional setting to control nr of digits to round}
}
\description{
Summarize a PRIM diversify result object
}
\author{
Jurian Baas
}
