% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{summary.prim.peel}
\alias{summary.prim.peel}
\title{Summarize a PRIM peeling result object}
\usage{
\method{summary}{prim.peel}(object, ..., round = TRUE, digits = 2)
}
\arguments{
\item{object}{An S3 object of class prim.peel}

\item{...}{Optional arguments to pass on}

\item{round}{Optional setting to disable rounding}

\item{digits}{Optional setting to control nr of digits to round}
}
\value{
Nothing, this function is called for its side-effects
}
\description{
Summarize a PRIM peeling result object
}
\author{
Jurian Baas
}
